- Moved dependencies to versions catalog - Moved AboutCard component to a separated file
